(Maint)|Add example project (part 1) (#6)
This adds a sample project, RelayExampleIOS, along with early documentation / walkthroughs. There's still plenty of documentation to add and refinement to be done, but I found this was a good stopping point for an initial PR.

The example project is a single-screen recipe-viewer. The content ultimately flows from a RecipeServiceType, which is the target of DDI for real-world demonstration within RelayExampleIOSUITests.