(Feature)|Add @Injected property wrapper for auto-injection (#17)
This takes advantage of `@propertyWrapper`, which was introduced in Swift 5.1 as the backbone to `@Binding` and other Apple-specific language features. Unlike `@Binding` and its counterparts, user-defined property wrappers are not limited to iOS 13+; they are instead bound by Swift version, as we would hope for.

The motivation behind this change is to make usage more declarative and easy-to-read. This also enforces initialization semantics across all usages, which prevents behavior inconsistencies between various components that use Relay for dependency injection.

The new property-wrapper is simply called `@Injected`, which is handled through the `Injected<V>` property wrapper class. The attribute can optionally be fed a scope, but defaults to the `.global` scope. The result: much cleaner, easier-to-read code:

```swift
// BEFORE

class FooClass {
    var barDependency: BarType = DependencyContainer.global.resolve()
    var bazDependency: BazType = DependencyContainer.container(for: .myCustomScope).resolve()
}

// AFTER

class FooClass {
    @Injected var barDependency: BarType
    @Injected(scope: .myCustomScope) var bazDependency: BazType
}
```